backend/requirements.txt
asynctest
bcrypt
fastapi
passlib
pytest
python-jose
python-multipart
tortoise-orm
uvicorn
requests